pointerevents/pointerup.html

	<!doctype html>
	<head>
	<meta name="viewport" content="width=device-width">
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	</head>

	<body onload="run()">
	<h2>PointerUp</h2>
	<h4>Test Description: This test checks if pointerup event triggers. Press mouse left button and release it over the black rectangle or tap it if you are using toucscreen.</h4>
	
	<p>		
	<div id="target0" style="background:black"></div>	
	<script>
	var eventTested = false;
	var checkTimer;
	var detected_pointertypes = {};

    setup({ explicit_done: true });
	add_completion_callback(showPointerTypes);

	function check_PointerEvent(event) {
	    test(function () {
	        assert_true(event instanceof PointerEvent,
				"event is a PointerEvent event");
	    }, event.type + " event is a PointerEvent event");
	}
	
	function run() {
		var target0 = document.getElementById("target0");
		var test_pointerup = async_test("pointerup event received");
		var pointerover_event;
		
		on_event(target0, "pointerover", function(event) {
			pointerover_event = event;
			detected_pointertypes[ event.pointerType ] = true;
		});

		on_event(target0, "pointerup", function (event) {
		    if (eventTested == false) {
		        check_PointerEvent(event);
		        document.getElementById('target0').style.background = 'green';
		        eventTested = true;

		        test_pointerup.step(function () {
		            assert_equals(event.pointerType, pointerover_event.pointerType,
					"pointerType is same for pointerover and pointerup");
		            assert_equals(event.isPrimary, pointerover_event.isPrimary,
					"isPrimary is same for pointerover and pointerup");
		        });

		        detected_pointertypes[event.pointerType] = true;

		        test_pointerup.done();
		        done();
		    }
		});
		
		on_event(target0, "mouseup", function(event) {
			checkTimer = setTimeout(checkResult, 1000);
		});
		
		on_event(target0, "touchend", function(event) {
			checkTimer = setTimeout(checkResult, 1000);
		});
	}
	
	function showPointerTypes() {
		var complete_notice = document.getElementById("complete-notice");
		var pointertype_log = document.getElementById("pointertype-log");
		var pointertypes = Object.keys(detected_pointertypes);
		pointertype_log.innerHTML = pointertypes.length ?
			pointertypes.join(",") : "(none)";
		complete_notice.style.display = "block";
	}
	
	function checkResult() {
		if(eventTested == false) {
		    document.getElementById('target0').style.background = 'red';
		    setup({ timeout: 0 });
		}
		clearTimeout(checkTimer);
	}
	</script>
	<style>
		div {
			margin: 0em;
			padding: 2em;
		}
		#complete-notice {
			background: #afa;
			border: 1px solid #0a0;
			display: none;
		}
		#pointertype-log {
			font-weight: bold;
		}
	</style>
	
	<h1>Pointer Events pointerup Tests</h1>
	<div id="complete-notice">
		<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
		<p>Would you like to <a href="#">run the tests again</a> with a different pointer type?</p>
	</div>
	<div id="log"></div>
	</body>
